gtable_combine {gridExtra}	R Documentation

Combine gtables based on row/column names.

Description

Combine gtables based on row/column names.

Usage

gtable_combine(..., along = 1L, join = "outer")

combine(..., along = 1L, join = "outer")

Arguments

...
gtables

along
dimension to align along, 1 = rows, 2 = cols.

join
when x and y have different names, how should the difference be resolved? inner keep names that appear in both, outer keep names that appear in either, left keep names from x, and right keep names from y.

[Package gridExtra version 2.3 Index]