tableGrob {gridExtra}	R Documentation

Graphical display of a textual table

Description

Create a gtable containing text grobs representing a character matrix.

Usage

tableGrob(d, rows = rownames(d), cols = colnames(d),
  theme = ttheme_default(), vp = NULL, ...)

grid.table(...)

ttheme_default(base_size = 12, base_colour = "black", base_family = "",
  parse = FALSE, padding = unit(c(4, 4), "mm"), ...)

ttheme_minimal(base_size = 12, base_colour = "black", base_family = "",
  parse = FALSE, padding = unit(c(4, 4), "mm"), ...)

Arguments

d
data.frame or matrix

rows
optional vector to specify row names

cols
optional vector to specify column names

theme
list of theme parameters

vp
optional viewport

...
further arguments to control the gtable

base_size
default font size

base_colour
default font colour

base_family
default font family

parse
logical, default behaviour for parsing text as plotmath

padding
length-2 unit vector specifying the horizontal and vertical padding of text within each cell

Value

A gtable.

Functions

tableGrob: return a grob

grid.table: draw a text table

ttheme_default: default theme for text tables

ttheme_minimal: minimalist theme for text tables

Examples

library(grid)
d <- head(iris, 3)
g <- tableGrob(d)
grid.newpage()
grid.draw(g)

[Package gridExtra version 2.3 Index]